chore(core): rename configurationBuilder
