+++ packages/core/__tests__/configurationBuilder/createConfiguration.test.ts
