+++ packages/core/src/configurationBuilder/configurationBuilder.ts
